move app/models/user_state_machine.rb -> app/services/users/state_machine.rb
